ee798/requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
Pillow==9.5.0
